Reinstate support for mangled Mesa.
Modified a bit from the original.  In this incarnation, OSMesa is
assumed to be mangled.  Mangled Mesa based on Mesa's fake glX is
probably broken, but that use case is not actually useful anyway.

So, now there are two supported modes: just build with GL only, as
per always, or build with both GL and mangled OSMesa.