Hack: claim OSMesa RW is direct.
This is a hack to get VTK to use OpenGL extensions more often.
The function is actually meant to detail whether a window is
accelerated by hardware or not, but in practice it gets used more
often to check for a broad class of OpenGL functionality that
historically has separated Mesa from other drivers.  In recent
years, the gap has dwindled, and furthermore few if any rendering
algorithms make use of much of anything above OpenGL 2.1 anyway.